Module libp2p_ping::handler

Structs

PingConfig

The configuration for outbound pings.

PingHandler

Protocol handler that handles pinging the remote at a regular period and answering ping queries.

Enums

PingFailure

An outbound ping failure.

PingSuccess

The successful result of processing an inbound or outbound ping.

Type Definitions

PingResult

The result of an inbound or outbound ping.