Struct libp2p_ping::Ping

pub struct Ping { /* fields omitted */ }

Ping is a NetworkBehaviour that responds to inbound pings and periodically sends outbound pings on every established connection.

See the crate root documentation for more information.

Implementations

impl Ping

pub fn new(config: PingConfig) -> Self

Creates a new Ping network behaviour with the given configuration.